terrible-inference/terrible_inference_server/inf_server/requirements.txt
# requirements.txt

# TODO sort this out later... 
# It would probably be in our best interest to have version numbers
Flask
gunicorn
psycopg2-binary
datetime
sqlite-utils
sentence
openai-whisper
numpy

